Fixes for penetration testing (from Plynt report).
	new file:   Configuration/DBScript/PenTestSQL.txt
	modified:   idas.chox.core/src/main/java/idas/chox/core/common/OrganisationType.java
	modified:   idas.chox.core/src/main/java/idas/chox/core/search/ClaimSearchCriteria.java
	modified:   idas.chox.data/src/main/java/idas/chox/data/services/WebUserUserRoleServiceImpl.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/ActionResponse.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/admin/AdminUserService.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/reports/AdminWeeklyOverviewReport.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/reports/BillingInsurerReport.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/reports/ClaimRejectedReport.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/reports/ExcelReportBuilder.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/reports/InsurerPaymentReport.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/reports/InvoiceReport.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/reports/InvoiceSavingSummaryReport.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/reports/InvoiceSummaryReport.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/reports/OverviewSummaryReport.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/security/ApplicationAccessibility.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/security/ReportAccessibility.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/AcknowledgeClaim.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/AssignOwner.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/AssignWorkgroup.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ClaimAwaitingCarHireInfo.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ClaimClosed.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ClaimOwnerAssigning.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ClaimPending.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ClaimReferToEng.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ClaimReferToFnol.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ClaimRegisterByFnol.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ClaimRejection.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ClaimRejectionAccept.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ClaimRejectionConstest.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ClaimReopen.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ClaimReviewByEng.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/InvoiceAccepted.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/InvoicePaymentLogged.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/InvoicePaymentReceived.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/InvoiceReferToCH.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/InvoiceReferToEng.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/InvoiceRejection.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/InvoiceRejectionAccept.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/InvoiceRejectionConstest.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/InvoiceResubmit.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/NewClaim.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/NewInvoice.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ResolveLiability.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/UpdateLiability.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/BaseAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/BillingAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/ClaimAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/ClaimActivityAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/InsurerAliasAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/InsurerAutomaticRoutingAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/InsurerBreBandAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/InsurerBreBandMappingAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/InsurerVehicleClassCeiling.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/ReportAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/SearchClaimAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/UserAccountAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/UserAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/UserroleAction.java
	new file:   idas.chox.web/src/main/java/idas/chox/web/security/CachingHeadersInterceptor.java
	modified:   idas.chox.web/src/main/resources/applicationContext.xml
	modified:   idas.chox.web/src/main/resources/struts.xml
	modified:   idas.chox.web/src/main/webapp/WEB-INF/decorators/main.jsp
	modified:   idas.chox.web/src/main/webapp/WEB-INF/web.xml
	modified:   idas.chox.web/src/main/webapp/jsp/adminPanel/p_user_detail_penal.jsp
	modified:   idas.chox.web/src/main/webapp/jsp/forget_password_request.jsp
	modified:   idas.chox.web/src/main/webapp/jsp/login.jsp
	new file:   idas.chox.web/src/main/webapp/jsp/p_jsonError.jsp
	modified:   idas.chox.web/src/main/webapp/jsp/p_user_account.jsp
	modified:   idas.chox.web/src/main/webapp/scripts/ajax.js

	modified:   idas.chox.service/src/main/java/idas/chox/service/security/ApplicationAccessibility.java

+++ idas.chox.web/src/main/java/idas/chox/web/actions/InsurerBreBandAction.java
@@ -1,16 +1,21 @@
 package idas.chox.web.actions;
 
+import java.util.ArrayList;
+import java.util.List;
+import net.sf.json.JSONArray;
 import com.opensymphony.xwork2.ModelDriven;
 import com.opensymphony.xwork2.Preparable;
+import org.springframework.security.annotation.Secured;
+import org.springframework.security.AccessDeniedException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import idas.chox.core.model.BreBand;
 import idas.chox.service.ActionResponse;
 import idas.chox.service.admin.AdminInsurerService;
 import idas.chox.web.viewdata.InsurerBreBandViewData;
-import java.util.ArrayList;
-import java.util.List;
-import net.sf.json.JSONArray;
 
 public class InsurerBreBandAction extends BaseAction implements ModelDriven<BreBand>, Preparable {
+    private static final Logger LOG = LoggerFactory.getLogger(InsurerBreBandAction.class);
 
     private String objectId;
     private int insurerId = -1;
@@ -95,9 +100,13 @@ public class InsurerBreBandAction extends BaseAction implements ModelDriven<BreB
         return SUCCESS;
     }
 
+    @Secured ({"ROLE_CHOX_ADMIN", "ROLE_INS_MNG"})
     public String updateInsurerBreBand() {
 
         try {
+            if ( getUserOrganisationType() == 3 || (getUserOrganisationType() == 2 && this.insurerId != getUserOrganisationId())) {
+                throw new AccessDeniedException("Trying to update an insurer BRE Band for an insurer that isn't mine (POSSIBLE HACK ATTEMPT)");
+            }
 
             ActionResponse response;
             response = adminInsurerService.updateInsurerBreBand(model, this.insurerId, getIsNew());
@@ -111,8 +120,12 @@ public class InsurerBreBandAction extends BaseAction implements ModelDriven<BreB
         return SUCCESS;
     }
 
+    @Secured ({"ROLE_CHOX_ADMIN", "ROLE_INS_MNG"})
     public String deleteInsurerBreBand() {
         try {
+            if ( getUserOrganisationType() == 3 || (getUserOrganisationType() == 2 && model.getInsurer().getId() != getUserOrganisationId())) {
+                throw new AccessDeniedException("Trying to delete an insurer BRE Band for an insurer that isn't mine (POSSIBLE HACK ATTEMPT)");
+            }
             
             ActionResponse response;
             response = adminInsurerService.deleteInsurerBreBand(model);
