+++ idas.chox.web/src/main/java/idas/chox/web/actions/UserroleAction.java
@@ -1,15 +1,24 @@
 package idas.chox.web.actions;
 
+import idas.chox.core.common.OrganisationType;
+import idas.chox.core.model.IdLookupItem;
+import idas.chox.core.model.WebUser;
 import idas.chox.core.model.WebUserRole;
 import idas.chox.core.model.WebUserUserRole;
 import idas.chox.service.ActionResponse;
 import idas.chox.service.admin.AdminUserService;
 import idas.chox.web.viewdata.UserroleViewData;
 import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 import net.sf.json.JSONArray;
+import org.springframework.security.annotation.Secured;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.security.AccessDeniedException;
 
 public class UserroleAction extends BaseAction {
+    private static final Logger LOG = LoggerFactory.getLogger(UserroleAction.class);
 
     private List<UserroleViewData> userroles;
     private int webUserId;
@@ -20,6 +29,7 @@ public class UserroleAction extends BaseAction {
     private AdminUserService adminUserService;
 
     public String doRenderActionPage() {
+        LOG.debug("doRenderActionPage() called for user: {}", webUserId);
         return SUCCESS;
     }
 
@@ -76,7 +86,7 @@ public class UserroleAction extends BaseAction {
     }
 
     public String getUseroles() {
-
+        LOG.debug("Getting user roles for user: {}", webUserId);
         try {
 
             userroles = new ArrayList<UserroleViewData>();
@@ -97,20 +107,41 @@ public class UserroleAction extends BaseAction {
         return SUCCESS;
     }
 
-    public List getAvailableUserroles() {
-        return adminUserService.getAvailableUserroles(organisationTypeId, webUserId);
+    public List<IdLookupItem> getAvailableUserroles() {
+          LOG.debug("Getting available user roles for user {} ({})", webUserId, organisationTypeId);
+      return adminUserService.getAvailableUserroles(organisationTypeId, webUserId);
     }
 
+// For some reason the following line causes the add/remove role panel to be displayed empty
+//    @Secured ({"ROLE_CHOX_ADMIN", "ROLE_INS_MNG", "ROLE_CHO_MNG"})
     public String addNewWebUserRoleMapping() {
+        LOG.debug("Adding user role '{}' to user '{}'", webUserRoleId, webUserId);
 
         if (webUserId > 0 && webUserRoleId > 0) {
 
             try {
-
+                // As Spring/ACEGI security is commented out (above), we'll check manually that we are an admin
+                if (!getIsAdmin()) {
+                    throw new AccessDeniedException("Non admin role trying to add a new role (POSSIBLE HACK ATTEMPT)");
+                }
+                // Need to check that that the user we are attaching the role to is one of our users
+                // this is to prevent parameter hacking
+                LOG.debug("Checking access to addNewWebUserRoleMapping for current user");
+                WebUser user = adminUserService.getUser(webUserId);
+                LOG.debug("User who me are adding roles to is: {} ('{}')", webUserId, user.getDisplayName());
+                if ((getUserOrganisationType() == 2 && (user.getInsurer() == null || getUserOrganisationId() != user.getInsurer().getId()))
+                         || (getUserOrganisationType() == 3 && (user.getChorganisation() == null || getUserOrganisationId() != user.getChorganisation().getId()))) {
+                    throw new AccessDeniedException("Trying to add a role to a user not of my organisation (POSSIBLE HACK ATTEMPT)");
+                }
+                // Check that the role is one we can add
+                if (!isRoleAvailable(webUserRoleId, OrganisationType.getOrganisationTypeId(user.getOrganisationType()))) {
+                    throw new AccessDeniedException("Trying to add a role not available (POSSIBLE HACK ATTEMPT)");
+                }
                 ActionResponse response = adminUserService.addNewWebUserRoleMapping(webUserId, webUserRoleId);
                 setActionResponse(response);
 
             } catch (Exception ex) {
+                LOG.debug("Handling exception: '{}'", ex.getMessage());
                 handleException(ex);
                 return ERROR;
             }
@@ -119,15 +150,45 @@ public class UserroleAction extends BaseAction {
         return SUCCESS;
     }
 
+    private boolean isRoleAvailable(int webUserRoleId, int orgType) {
+        LOG.debug("Is role {} available to this user?", webUserRoleId);
+        List<IdLookupItem> availableRoles = adminUserService.getAvailableUserroles(orgType, webUserId);
+        LOG.debug("We have {} roles available:", availableRoles.size());
+        for (Iterator<IdLookupItem> i = availableRoles.iterator(); i.hasNext( ); ) {
+            IdLookupItem lu = i.next();
+            LOG.debug("Role available: {} - '{}'", lu.getId(), lu.getName());
+            if (lu.getId() == webUserRoleId)
+                return true;
+        }
+        return false;
+    }
+
+// For some reason the following line causes the add/remove role panel to be displayed empty
+//    @Secured ({"ROLE_CHOX_ADMIN", "ROLE_INS_MNG", "ROLE_CHO_MNG"})
     public String removeWebUserRoleMapping() {
+        LOG.debug("Removing user role '{}' to user '{}'", webUserRoleId, webUserId);
 
         if (this.webUserUserRoleId > 0) {
 
             try {
-                ActionResponse response = adminUserService.deleteWebUserRoleMapping(this.webUserUserRoleId);
-                setActionResponse(response);
-
+                // As Spring/ACEGI security is commented out (above), we'll check manually that we are an admin
+                if (!getIsAdmin()) {
+                    throw new AccessDeniedException("Non admin role trying to remove a new role (POSSIBLE HACK ATTEMPT)");
+                }
+                 // Need to check that that the user we are attaching the role to is one of our users
+                 // this is to prevent parameter hacking
+                 LOG.debug("Checking access to removeNewWebUserRoleMapping for current user");
+                 WebUser user = adminUserService.getUser(webUserId);
+                 LOG.debug("User who me are removing roles from is: {} ('{}')", webUserId, user.getDisplayName());
+                 if ((getUserOrganisationType() == 2 && (user.getInsurer() == null || getUserOrganisationId() != user.getInsurer().getId()))
+                         || (getUserOrganisationType() == 3 && (user.getChorganisation() == null || getUserOrganisationId() != user.getChorganisation().getId()))) {
+                     LOG.debug("Throwing AccessDeniedException");
+                     throw new AccessDeniedException("Trying to remove a role to a user not of my organisation (POSSIBLE HACK ATTEMPT)");
+                 }
+                 ActionResponse response = adminUserService.deleteWebUserRoleMapping(this.webUserUserRoleId);
+                 setActionResponse(response);
             } catch (Exception ex) {
+                LOG.debug("Handling exception: '{}'", ex.getMessage());
                 handleException(ex);
                 return ERROR;
             }
